Backend/requirements.txt
blinker==1.9.0
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.0.5       # <- add this
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mongoengine==0.29.1
PyJWT==2.10.1
pymongo==4.10.1
Werkzeug==3.1.3
psycopg2-binary
python-dotenv
redis
minio
reportlab
flask-socketio